Process the org.freedesktop.DBus.Introspectable and org.freedesktop.DBus.Peer interfaces in their own modules
